% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterNodeSelection.R
\name{ClusterNodeSelection}
\alias{ClusterNodeSelection}
\title{Test the validity of a subcluster directly under a given cluster}
\usage{
ClusterNodeSelection(cluster, testPos, alpha, data, BootData)
}
\arguments{
\item{cluster}{the main cluster}

\item{testPos}{the position of the subcluster to test, directly under the mian cluster}

\item{alpha}{the confidence level for the tests}

\item{data}{the underlying data}

\item{BootData}{the datafram of bootstrap samples of spearman rho, with columsn
named "(i,j)", where "i" and "j" are different leave}
}
\value{
The main cluster, with or without the node un der test, wether the
hypothesis can be rejected or not.
}
\description{
Test the validity of a subcluster directly under a given cluster
}
\author{
Simon-Pierre Gadoury
}
