% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLeaves.R
\name{GetLeaves}
\alias{GetLeaves}
\title{Leaves extractions}
\usage{
GetLeaves(cluster)
}
\arguments{
\item{cluster}{a cluster (of the form provided by the function hclust2tree)}
}
\value{
Vector of leaves.
}
\description{
Extract the leaves (nodes in the bottom of the hierarchy) under
(not necessarily directly under) a cluster.
}
\author{
Simon-Pierre Gadoury
}
