% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPairs.R
\name{GetPairs}
\alias{GetPairs}
\title{Get pairs of leaves from different branches of a cluster}
\usage{
GetPairs(cluster, pos)
}
\arguments{
\item{cluster}{main cluster.}

\item{pos}{position of the subcluster directly under the main cluster.
This cluster is the one every other cluster will be paired with.}
}
\value{
Matrice of pairs of leaves.
}
\description{
Get pairs of leaves from different branches of a cluster
}
\author{
Simon-Pierre Gadoury
}
