% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifySpearman.R
\name{ModifySpearman}
\alias{ModifySpearman}
\title{Modifed Spearman's rho matrix}
\usage{
ModifySpearman(Tree, EmpSpearmanMatrix)
}
\arguments{
\item{Tree}{tree structure to use for the grouping}

\item{EmpSpearmanMatrix}{initial empirical Spearman's rho matrix to modify}
}
\description{
Modifies, according to the tree structure provided, the empirical Spearman's rho matrix.
}
\author{
Simon-Pierre Gadoury
}
