% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2graphdf.R
\name{tree2graphdf}
\alias{tree2graphdf}
\title{Tree to graph dataframe}
\usage{
tree2graphdf(tree, plot = TRUE)
}
\arguments{
\item{tree}{the tree under consideration}

\item{plot}{should a basic plot of the tree be shown ?}
}
\value{
a graph.data.frame object
}
\description{
Tree to graph dataframe
}
\author{
Simon-Pierre Gadoury
}
