% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2plot.R
\name{tree2plot}
\alias{tree2plot}
\title{Plot of a tree structure}
\usage{
tree2plot(tree, plot = TRUE, ...)
}
\arguments{
\item{tree}{the tree under consideration (of the form provided by the function hclust2tree)}

\item{plot}{show a basic plot the tree ?}

\item{...}{extra parameters passed to igraph.plot}
}
\value{
a plot or graph.data.frame object
}
\description{
Plot of a tree structure
}
\author{
Simon-Pierre Gadoury
}
