% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyTree.R
\name{VerifyTree}
\alias{VerifyTree}
\title{Verify tree structure}
\usage{
VerifyTree(tree, data, alpha = 0.95, nboot = 500)
}
\arguments{
\item{tree}{the main tree}

\item{data}{the data used to construct the tree}

\item{alpha}{the confidence level for the tests}

\item{nboot}{the number of bootstrap samples to use}
}
\value{
The tree, modified, according to the results of the tests.
}
\description{
Verify tree structure
}
\author{
Simon-Pierre Gadoury
}
