% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterNodeSelection.R
\name{ClusterNodeSelection}
\alias{ClusterNodeSelection}
\title{Subcluster significance test}
\usage{
ClusterNodeSelection(cluster, testPos, alpha, data, BootData)
}
\arguments{
\item{cluster}{the main cluster (of the form provided by the function hclust2tree)}

\item{testPos}{the position of the subcluster to test, directly under the main cluster}

\item{alpha}{the confidence level for the tests}

\item{data}{the underlying data}

\item{BootData}{the dataframe of bootstrap samples of Spearman rho, with columns
named "(i,j)", where "i" and "j" are different leaves}
}
\value{
The main cluster, with or without the node under test, wether the
hypothesis can be rejected or not.
}
\description{
Test the significance of a subcluster directly under a given cluster.
If it is not significant, the subcluster is eliminated and
its element are merged in the main cluster. Bootstrap samples of the underlying
data needs to be provided, as well as the structure under consideration.
}
\details{
The hypothesis testing is made with the matrix of Spearman's rho
for a given dataset, see \insertCite{gaisser2010testing}{erhcv}.
}
\references{
\insertRef{gaisser2010testing}{erhcv}
}
\author{
Simon-Pierre Gadoury
}
