% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliminateCluster.R
\name{EliminateCluster}
\alias{EliminateCluster}
\title{Subcluster elimination}
\usage{
EliminateCluster(cluster, pos)
}
\arguments{
\item{cluster}{the cluster under consideration (of the form provided by the function hclust2tree)}

\item{pos}{the position of the subcluster, directly under the main cluster, to eliminate}
}
\value{
The main cluster, without the subcluster that was eliminated.
}
\description{
Eliminate a subcluster and merge its elements to the main cluster.
}
\author{
Simon-Pierre Gadoury
}
