% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliminateCluster.R
\name{EliminateCluster}
\alias{EliminateCluster}
\title{Eliminate a subcluster from a cluster}
\usage{
EliminateCluster(cluster, pos)
}
\arguments{
\item{cluster}{the cluster under consideration}

\item{pos}{the position of the subcluster, directly under the main cluster, to eliminate}
}
\value{
The main cluster, without the subcluster under consideration.
}
\description{
Eliminate a subcluster from a cluster
}
\author{
Simon-Pierre Gadoury
}
