% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLeaves.R
\name{GetLeaves}
\alias{GetLeaves}
\title{Get a vector of leaves under a given cluster}
\usage{
GetLeaves(cluster)
}
\arguments{
\item{cluster}{a cluster}
}
\value{
Vector of leaves under the cluster.
}
\description{
Get a vector of leaves under a given cluster
}
\author{
Simon-Pierre Gadoury
}
