% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPairs.R
\name{GetPairs}
\alias{GetPairs}
\title{Obtain pairs of leaves}
\usage{
GetPairs(cluster, pos)
}
\arguments{
\item{cluster}{main cluster (of the form provided by the function hclust2tree)}

\item{pos}{position of the subcluster directly under the main cluster.
This subcluster is the one every other subcluster will be paired with.}
}
\value{
Matrix of pairs of leaves.
}
\description{
Find all the pairs of leaves between
a specific subcluster and all the other subclusters
under the main cluster. Pairs between the subclusters
of the specific subcluster are also given.
}
\author{
Simon-Pierre Gadoury
}
