% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust2tree.R
\name{hclust2tree}
\alias{hclust2tree}
\title{Transform a hclust object into a tree (list)}
\usage{
hclust2tree(clustering)
}
\arguments{
\item{clustering}{hclust object}
}
\description{
Transform a hclust object into a tree (list)
}
\author{
Simon-Pierre Gadoury
}
