% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust2tree.R
\name{hclust2tree}
\alias{hclust2tree}
\title{hclust object transformation}
\usage{
hclust2tree(clustering)
}
\arguments{
\item{clustering}{hclust object}
}
\description{
Transforms a hclust object into a tree (cluster) used
in other functions of this package.
}
\author{
Simon-Pierre Gadoury
}
