% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2plot.R
\name{tree2plot}
\alias{tree2plot}
\title{Plot of a tree structure}
\usage{
tree2plot(tree)
}
\arguments{
\item{tree}{the tree under consideration of the form provided by the function hclust2tree)}
}
\value{
a plot or graph.data.frame object
}
\description{
Provide a plot or a graph.data.frame object for a given tree structure.
}
\author{
Simon-Pierre Gadoury
}
