% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2plot.R
\name{tree2plot}
\alias{tree2plot}
\title{Plot of a tree structure}
\usage{
tree2plot(tree, structure = FALSE)
}
\arguments{
\item{tree}{the tree under consideration (of the form provided by the function hclust2tree)}

\item{structure}{logical. Should the data.tree structure be returned rather than the plot?}
}
\value{
a plot or data.tree object
}
\description{
Provide the plot or the data.tree object for a given tree structure.
}
\author{
Simon-Pierre Gadoury
}
